--- src/pages/SwapDeprecated/direct.tsx
@@ -1,563 +1,0 @@
-import { CurrencyAmount, JSBI, Token, Trade } from '@sushiswap/sdk'
-import UnsupportedCurrencyFooter from 'components/swap/UnsupportedCurrencyFooter'
-import { useIsTransactionUnsupported } from 'hooks/Trades'
-import React, { useCallback, useContext, useEffect, useMemo, useState } from 'react'
-import { ArrowDown } from 'react-feather'
-
-import { Text } from 'rebass'
-import { useTheme } from 'styled-components'
-import AddressInputPanel from '../../components/AddressInputPanel'
-import { ButtonConfirmed, ButtonError, ButtonLight, ButtonPrimary } from '../../components/ButtonLegacy'
-import Card, { GreyCard } from '../../components/CardLegacy'
-import Column, { AutoColumn } from '../../components/Column'
-import CurrencyInputPanel from '../../components/CurrencyInputPanel'
-import Loader from '../../components/Loader'
-import ProgressSteps from '../../components/ProgressSteps'
-import { AutoRow, RowBetween } from '../../components/Row'
-import { AdvancedSwapDetails } from '../../components/swap/AdvancedSwapDetails'
-import confirmPriceImpactWithoutFee from '../../components/swap/confirmPriceImpactWithoutFee'
-import ConfirmSwapModal from '../../components/swap/ConfirmSwapModal'
-import { ArrowWrapper, BottomGrouping, SwapCallbackError, Wrapper } from '../../components/swap/styleds'
-import SwapHeader from '../../components/ExchangeHeader'
-import TradePrice from '../../components/swap/TradePrice'
-import { INITIAL_ALLOWED_SLIPPAGE } from '../../constants'
-import { useActiveWeb3React } from '../../hooks/useActiveWeb3React'
-import { useAllTokens, useCurrency } from '../../hooks/Tokens'
-import { ApprovalState, useApproveCallbackFromTrade } from '../../hooks/useApproveCallback'
-import useENSAddress from '../../hooks/useENSAddress'
-import { useSwapCallback } from '../../hooks/useSwapCallback'
-import useWrapCallback, { WrapType } from '../../hooks/useWrapCallback'
-import { useToggleSettingsMenu, useWalletModalToggle } from '../../state/application/hooks'
-import { Field } from '../../state/swap/actions'
-import {
-    useDefaultsFromURLSearch,
-    useDerivedSwapInfo,
-    useSwapActionHandlers,
-    useSwapState
-} from '../../state/swap/hooks'
-import { useExpertModeManager, useUserSingleHopOnly, useUserSlippageTolerance } from '../../state/user/hooks'
-import { LinkStyledButton, TYPE } from '../../theme'
-import { maxAmountSpend } from '../../utils/maxAmountSpend'
-import { computeTradePriceBreakdown, warningSeverity } from '../../utils/prices'
-import { ClickableText } from '../Pool/styleds'
-import { t } from '@lingui/macro'
-import { useLingui } from '@lingui/react'
-
-export default function Swap() {
-    const { i18n } = useLingui()
-    const loadedUrlParams = useDefaultsFromURLSearch()
-
-    // token warning stuff
-    const [loadedInputCurrency, loadedOutputCurrency] = [
-        useCurrency(loadedUrlParams?.inputCurrencyId),
-        useCurrency(loadedUrlParams?.outputCurrencyId)
-    ]
-    const [dismissTokenWarning, setDismissTokenWarning] = useState<boolean>(false)
-    const urlLoadedTokens: Token[] = useMemo(
-        () => [loadedInputCurrency, loadedOutputCurrency]?.filter((c): c is Token => c instanceof Token) ?? [],
-        [loadedInputCurrency, loadedOutputCurrency]
-    )
-    const handleConfirmTokenWarning = useCallback(() => {
-        setDismissTokenWarning(true)
-    }, [])
-
-    // dismiss warning if all imported tokens are in active lists
-    const defaultTokens = useAllTokens()
-    const importTokensNotInDefault =
-        urlLoadedTokens &&
-        urlLoadedTokens.filter((token: Token) => {
-            return !Boolean(token.address in defaultTokens)
-        })
-
-    const { account, chainId } = useActiveWeb3React()
-    const theme = useTheme()
-
-    // toggle wallet when disconnected
-    const toggleWalletModal = useWalletModalToggle()
-
-    // for expert mode
-    const toggleSettings = useToggleSettingsMenu()
-    const [isExpertMode] = useExpertModeManager()
-
-    // get custom setting values for user
-    const [allowedSlippage] = useUserSlippageTolerance()
-
-    // swap state
-    const { independentField, typedValue, recipient } = useSwapState()
-    const { v2Trade, currencyBalances, parsedAmount, currencies, inputError: swapInputError } = useDerivedSwapInfo()
-    const { wrapType, execute: onWrap, inputError: wrapInputError } = useWrapCallback(
-        currencies[Field.INPUT],
-        currencies[Field.OUTPUT],
-        typedValue
-    )
-    const showWrap: boolean = wrapType !== WrapType.NOT_APPLICABLE
-    const { address: recipientAddress } = useENSAddress(recipient)
-
-    const trade = showWrap ? undefined : v2Trade
-
-    const parsedAmounts = showWrap
-        ? {
-            [Field.INPUT]: parsedAmount,
-            [Field.OUTPUT]: parsedAmount
-        }
-        : {
-            [Field.INPUT]: independentField === Field.INPUT ? parsedAmount : trade?.inputAmount,
-            [Field.OUTPUT]: independentField === Field.OUTPUT ? parsedAmount : trade?.outputAmount
-        }
-
-    const { onSwitchTokens, onCurrencySelection, onUserInput, onChangeRecipient } = useSwapActionHandlers()
-    const isValid = !swapInputError
-    const dependentField: Field = independentField === Field.INPUT ? Field.OUTPUT : Field.INPUT
-
-    const handleTypeInput = useCallback(
-        (value: string) => {
-            onUserInput(Field.INPUT, value)
-        },
-        [onUserInput]
-    )
-    const handleTypeOutput = useCallback(
-        (value: string) => {
-            onUserInput(Field.OUTPUT, value)
-        },
-        [onUserInput]
-    )
-
-    // modal and loading
-    const [{ showConfirm, tradeToConfirm, swapErrorMessage, attemptingTxn, txHash }, setSwapState] = useState<{
-        showConfirm: boolean
-        tradeToConfirm: Trade | undefined
-        attemptingTxn: boolean
-        swapErrorMessage: string | undefined
-        txHash: string | undefined
-    }>({
-        showConfirm: false,
-        tradeToConfirm: undefined,
-        attemptingTxn: false,
-        swapErrorMessage: undefined,
-        txHash: undefined
-    })
-
-    const formattedAmounts = {
-        [independentField]: typedValue,
-        [dependentField]: showWrap
-            ? parsedAmounts[independentField]?.toExact() ?? ''
-            : parsedAmounts[dependentField]?.toSignificant(6) ?? ''
-    }
-
-    const route = trade?.route
-    const userHasSpecifiedInputOutput = Boolean(
-        currencies[Field.INPUT] &&
-        currencies[Field.OUTPUT] &&
-        parsedAmounts[independentField]?.greaterThan(JSBI.BigInt(0))
-    )
-    const noRoute = !route
-
-    // check whether the user has approved the router on the input token
-    const [approval, approveCallback] = useApproveCallbackFromTrade(trade, allowedSlippage)
-
-    // check if user has gone through approval process, used to show two step buttons, reset on token change
-    const [approvalSubmitted, setApprovalSubmitted] = useState<boolean>(false)
-
-    // mark when a user has submitted an approval, reset onTokenSelection for input field
-    useEffect(() => {
-        if (approval === ApprovalState.PENDING) {
-            setApprovalSubmitted(true)
-        }
-    }, [approval, approvalSubmitted])
-
-    const maxAmountInput: CurrencyAmount | undefined = maxAmountSpend(currencyBalances[Field.INPUT])
-    const atMaxAmountInput = Boolean(maxAmountInput && parsedAmounts[Field.INPUT]?.equalTo(maxAmountInput))
-
-    // the callback to execute the swap
-    const { callback: swapCallback, error: swapCallbackError } = useSwapCallback(trade, allowedSlippage, recipient)
-
-    const { priceImpactWithoutFee } = computeTradePriceBreakdown(trade)
-
-    const [singleHopOnly] = useUserSingleHopOnly()
-
-    const handleSwap = useCallback(() => {
-        if (priceImpactWithoutFee && !confirmPriceImpactWithoutFee(priceImpactWithoutFee)) {
-            return
-        }
-        if (!swapCallback) {
-            return
-        }
-        setSwapState({
-            attemptingTxn: true,
-            tradeToConfirm,
-            showConfirm,
-            swapErrorMessage: undefined,
-            txHash: undefined
-        })
-        swapCallback()
-            .then(hash => {
-                setSwapState({
-                    attemptingTxn: false,
-                    tradeToConfirm,
-                    showConfirm,
-                    swapErrorMessage: undefined,
-                    txHash: hash
-                })
-            })
-            .catch(error => {
-                setSwapState({
-                    attemptingTxn: false,
-                    tradeToConfirm,
-                    showConfirm,
-                    swapErrorMessage: error.message,
-                    txHash: undefined
-                })
-            })
-    }, [
-        priceImpactWithoutFee,
-        swapCallback,
-        tradeToConfirm,
-        showConfirm,
-        recipient,
-        recipientAddress,
-        account,
-        trade,
-        singleHopOnly
-    ])
-
-    // errors
-    const [showInverted, setShowInverted] = useState<boolean>(false)
-
-    // warnings on slippage
-    const priceImpactSeverity = warningSeverity(priceImpactWithoutFee)
-
-    // show approve flow when: no error on inputs, not approved or pending, or approved in current session
-    // never show if price impact is above threshold in non expert mode
-    const showApproveFlow =
-        !swapInputError &&
-        (approval === ApprovalState.NOT_APPROVED ||
-            approval === ApprovalState.PENDING ||
-            (approvalSubmitted && approval === ApprovalState.APPROVED)) &&
-        !(priceImpactSeverity > 3 && !isExpertMode)
-
-    const handleConfirmDismiss = useCallback(() => {
-        setSwapState({ showConfirm: false, tradeToConfirm, attemptingTxn, swapErrorMessage, txHash })
-        // if there was a tx hash, we want to clear the input
-        if (txHash) {
-            onUserInput(Field.INPUT, '')
-        }
-    }, [attemptingTxn, onUserInput, swapErrorMessage, tradeToConfirm, txHash])
-
-    const handleAcceptChanges = useCallback(() => {
-        setSwapState({ tradeToConfirm: trade, swapErrorMessage, txHash, attemptingTxn, showConfirm })
-    }, [attemptingTxn, showConfirm, swapErrorMessage, trade, txHash])
-
-    const handleInputSelect = useCallback(
-        inputCurrency => {
-            setApprovalSubmitted(false) // reset 2 step UI for approvals
-            onCurrencySelection(Field.INPUT, inputCurrency)
-        },
-        [onCurrencySelection]
-    )
-
-    const handleMaxInput = useCallback(() => {
-        maxAmountInput && onUserInput(Field.INPUT, maxAmountInput.toExact())
-    }, [maxAmountInput, onUserInput])
-
-    const handleOutputSelect = useCallback(outputCurrency => onCurrencySelection(Field.OUTPUT, outputCurrency), [
-        onCurrencySelection
-    ])
-
-    const swapIsUnsupported = useIsTransactionUnsupported(currencies?.INPUT, currencies?.OUTPUT)
-
-    return (
-        <>
-            <div className="relative w-full max-w-lg rounded ark-900">
-                <SwapHeader />
-                <Wrapper id="swap-page">
-                    <ConfirmSwapModal
-                        isOpen={showConfirm}
-                        trade={trade}
-                        originalTrade={tradeToConfirm}
-                        onAcceptChanges={handleAcceptChanges}
-                        attemptingTxn={attemptingTxn}
-                        txHash={txHash}
-                        recipient={recipient}
-                        allowedSlippage={allowedSlippage}
-                        onConfirm={handleSwap}
-                        swapErrorMessage={swapErrorMessage}
-                        onDismiss={handleConfirmDismiss}
-                    />
-                    {/* <AutoColumn gap={isExpertMode ? 'md' : 'none'} style={{ paddingBottom: '1rem' }}> */}
-                    <AutoColumn gap={isExpertMode ? 'md' : '3px'}>
-                        <CurrencyInputPanel
-                            label={
-                                independentField === Field.OUTPUT && !showWrap && trade
-                                    ? i18n._(t`From (estimated)`)
-                                    : i18n._(t`From`)
-                            }
-                            value={formattedAmounts[Field.INPUT]}
-                            showMaxButton={!atMaxAmountInput}
-                            currency={currencies[Field.INPUT]}
-                            onUserInput={handleTypeInput}
-                            onMax={handleMaxInput}
-                            onCurrencySelect={handleInputSelect}
-                            otherCurrency={currencies[Field.OUTPUT]}
-                            id="swap-currency-input"
-                            cornerRadiusBottomNone={isExpertMode ? false : true}
-                        //containerBackground={'#101b31'}
-                        />
-                        {isExpertMode && !showWrap && (
-                            <AutoColumn justify="space-between">
-                                <AutoRow
-                                    justify={isExpertMode ? 'space-between' : 'center'}
-                                    style={{ padding: '0 1rem' }}
-                                >
-                                    <ArrowWrapper clickable>
-                                        <ArrowDown
-                                            size="16"
-                                            onClick={() => {
-                                                setApprovalSubmitted(false) // reset 2 step UI for approvals
-                                                onSwitchTokens()
-                                            }}
-                                            color={
-                                                currencies[Field.INPUT] && currencies[Field.OUTPUT]
-                                                    ? theme.primary1
-                                                    : theme.text2
-                                            }
-                                        />
-                                    </ArrowWrapper>
-                                    {recipient === null && !showWrap && isExpertMode ? (
-                                        <LinkStyledButton
-                                            id="add-recipient-button"
-                                            onClick={() => onChangeRecipient('')}
-                                        >
-                                            {i18n._(t`+ Add a send (optional)`)}
-                                        </LinkStyledButton>
-                                    ) : null}
-                                </AutoRow>
-                            </AutoColumn>
-                        )}
-                        {!isExpertMode && (
-                            <div style={{ position: 'relative', zIndex: 2 }}>
-                                <div
-                                    style={{
-                                        position: 'absolute',
-                                        width: '100%',
-                                        cursor: 'pointer',
-                                        marginTop: '-10px'
-                                    }}
-                                    onClick={() => {
-                                        setApprovalSubmitted(false) // reset 2 step UI for approvals
-                                        onSwitchTokens()
-                                    }}
-                                >
-                                    <svg
-                                        style={{ height: '1.4rem', width: '100%', margin: '0 auto' }}
-                                        fill="none"
-                                        stroke="currentColor"
-                                        viewBox="0 0 24 24"
-                                        xmlns="http://www.w3.org/2000/svg"
-                                    >
-                                        <path
-                                            strokeLinecap="round"
-                                            strokeLinejoin="round"
-                                            strokeWidth={2}
-                                            d="M7 16V4m0 0L3 8m4-4l4 4m6 0v12m0 0l4-4m-4 4l-4-4"
-                                        />
-                                    </svg>
-                                </div>
-                            </div>
-                        )}
-                        <CurrencyInputPanel
-                            value={formattedAmounts[Field.OUTPUT]}
-                            onUserInput={handleTypeOutput}
-                            label={
-                                independentField === Field.INPUT && !showWrap && trade
-                                    ? i18n._(t`To (estimated)`)
-                                    : i18n._(t`To`)
-                            }
-                            showMaxButton={false}
-                            currency={currencies[Field.OUTPUT]}
-                            onCurrencySelect={handleOutputSelect}
-                            otherCurrency={currencies[Field.INPUT]}
-                            id="swap-currency-output"
-                            cornerRadiusTopNone={isExpertMode ? false : true}
-                        //containerBackground={'#16182b'}
-                        />
-
-                        {recipient !== null && !showWrap ? (
-                            <>
-                                <AutoRow justify="space-between" style={{ padding: '0 1rem' }}>
-                                    <ArrowWrapper clickable={false}>
-                                        <ArrowDown size="16" color={theme.text2} />
-                                    </ArrowWrapper>
-                                    <LinkStyledButton
-                                        id="remove-recipient-button"
-                                        onClick={() => onChangeRecipient(null)}
-                                    >
-                                        {i18n._(t`- Remove send`)}
-                                    </LinkStyledButton>
-                                </AutoRow>
-                                <AddressInputPanel id="recipient" value={recipient} onChange={onChangeRecipient} />
-                            </>
-                        ) : null}
-                    </AutoColumn>
-                    <BottomGrouping style={{ paddingBottom: '1rem' }}>
-                        {swapIsUnsupported ? (
-                            <ButtonPrimary disabled={true}>
-                                <TYPE.main mb="4px">{i18n._(t`Unsupported Asset`)}</TYPE.main>
-                            </ButtonPrimary>
-                        ) : !account ? (
-                            <ButtonLight onClick={toggleWalletModal}>{i18n._(t`Connect Wallet`)}</ButtonLight>
-                        ) : showWrap ? (
-                            <ButtonPrimary disabled={Boolean(wrapInputError)} onClick={onWrap}>
-                                {wrapInputError ??
-                                    (wrapType === WrapType.WRAP
-                                        ? i18n._(t`Wrap`)
-                                        : wrapType === WrapType.UNWRAP
-                                            ? i18n._(t`Unwrap`)
-                                            : null)}
-                            </ButtonPrimary>
-                        ) : noRoute && userHasSpecifiedInputOutput ? (
-                            <GreyCard style={{ textAlign: 'center' }}>
-                                <TYPE.main mb="4px">{i18n._(t`Insufficient liquidity for this trade`)}</TYPE.main>
-                                {singleHopOnly && (
-                                    <TYPE.main mb="4px">{i18n._(t`Try enabling multi-hop trades`)}</TYPE.main>
-                                )}
-                            </GreyCard>
-                        ) : showApproveFlow ? (
-                            <RowBetween>
-                                <ButtonConfirmed
-                                    onClick={approveCallback}
-                                    disabled={approval !== ApprovalState.NOT_APPROVED || approvalSubmitted}
-                                    width="48%"
-                                    altDisabledStyle={approval === ApprovalState.PENDING} // show solid button while waiting
-                                    confirmed={approval === ApprovalState.APPROVED}
-                                >
-                                    {approval === ApprovalState.PENDING ? (
-                                        <AutoRow gap="6px" justify="center">
-                                            {i18n._(t`Approving`)} <Loader stroke="white" />
-                                        </AutoRow>
-                                    ) : approvalSubmitted && approval === ApprovalState.APPROVED ? (
-                                        i18n._(t`Approved`)
-                                    ) : (
-                                        i18n._(t`Approve ${currencies[Field.INPUT]?.getSymbol(chainId)}`)
-                                    )}
-                                </ButtonConfirmed>
-                                <ButtonError
-                                    onClick={() => {
-                                        if (isExpertMode) {
-                                            handleSwap()
-                                        } else {
-                                            setSwapState({
-                                                tradeToConfirm: trade,
-                                                attemptingTxn: false,
-                                                swapErrorMessage: undefined,
-                                                showConfirm: true,
-                                                txHash: undefined
-                                            })
-                                        }
-                                    }}
-                                    width="48%"
-                                    id="swap-button"
-                                    disabled={
-                                        !isValid ||
-                                        approval !== ApprovalState.APPROVED ||
-                                        (priceImpactSeverity > 3 && !isExpertMode)
-                                    }
-                                    error={isValid && priceImpactSeverity > 2}
-                                >
-                                    <Text fontSize={16} fontWeight={500}>
-                                        {priceImpactSeverity > 3 && !isExpertMode
-                                            ? i18n._(t`Price Impact High`)
-                                            : priceImpactSeverity > 2
-                                                ? i18n._(t`Swap Anyway`)
-                                                : i18n._(t`Swap`)}
-                                    </Text>
-                                </ButtonError>
-                            </RowBetween>
-                        ) : (
-                            <ButtonError
-                                onClick={() => {
-                                    if (isExpertMode) {
-                                        handleSwap()
-                                    } else {
-                                        setSwapState({
-                                            tradeToConfirm: trade,
-                                            attemptingTxn: false,
-                                            swapErrorMessage: undefined,
-                                            showConfirm: true,
-                                            txHash: undefined
-                                        })
-                                    }
-                                }}
-                                id="swap-button"
-                                disabled={!isValid || (priceImpactSeverity > 3 && !isExpertMode) || !!swapCallbackError}
-                                error={isValid && priceImpactSeverity > 2 && !swapCallbackError}
-                            >
-                                <Text fontSize={20} fontWeight={500}>
-                                    {swapInputError
-                                        ? swapInputError
-                                        : priceImpactSeverity > 3 && !isExpertMode
-                                            ? i18n._(t`Price Impact Too High`)
-                                            : priceImpactSeverity > 2
-                                                ? i18n._(t`Swap Anyway`)
-                                                : i18n._(t`Swap`)}
-                                </Text>
-                            </ButtonError>
-                        )}
-                        {showApproveFlow && (
-                            <Column style={{ marginTop: '1rem' }}>
-                                <ProgressSteps steps={[approval === ApprovalState.APPROVED]} />
-                            </Column>
-                        )}
-                        {isExpertMode && swapErrorMessage ? <SwapCallbackError error={swapErrorMessage} /> : null}
-                    </BottomGrouping>
-                    <AutoColumn>
-                        {showWrap ? null : (
-                            <Card padding={showWrap ? '.25rem 1rem 0 1rem' : '0px'} borderRadius={'20px'}>
-                                <AutoColumn gap="8px" style={{ padding: '0 16px' }}>
-                                    {Boolean(trade) && (
-                                        <RowBetween align="center">
-                                            <Text fontWeight={500} fontSize={14} color={theme.text3}>
-                                                {i18n._(t`Price`)}
-                                            </Text>
-                                            <TradePrice
-                                                price={trade?.executionPrice}
-                                                showInverted={showInverted}
-                                                setShowInverted={setShowInverted}
-                                            />
-                                        </RowBetween>
-                                    )}
-                                    {allowedSlippage !== INITIAL_ALLOWED_SLIPPAGE && (
-                                        <RowBetween align="center">
-                                            <ClickableText
-                                                fontWeight={500}
-                                                fontSize={14}
-                                                color={theme.text2}
-                                                onClick={toggleSettings}
-                                            >
-                                                {i18n._(t`Slippage Tolerance`)}
-                                            </ClickableText>
-                                            <ClickableText
-                                                fontWeight={500}
-                                                fontSize={14}
-                                                color={theme.text2}
-                                                onClick={toggleSettings}
-                                            >
-                                                {allowedSlippage / 100}%
-                                            </ClickableText>
-                                        </RowBetween>
-                                    )}
-                                </AutoColumn>
-                            </Card>
-                        )}
-                        <AdvancedSwapDetails trade={trade} />
-                    </AutoColumn>
-                </Wrapper>
-            </div>
-            {!swapIsUnsupported ? null : (
-                <UnsupportedCurrencyFooter
-                    show={swapIsUnsupported}
-                    currencies={[currencies.INPUT, currencies.OUTPUT]}
-                />
-            )}
-        </>
-    )
-}
